+++ packages/get_hooked/lib/listenables.dart
@@ -4,6 +4,7 @@ export 'package:flutter/foundation.dart' show ValueListenable;
 
 export 'listenables/async.dart';
 export 'listenables/default_animation_style.dart';
+export 'listenables/media_query.dart';
 export 'listenables/proxy.dart';
 export 'listenables/styled_animation.dart';
 export 'listenables/value_animation.dart';